+++ packages/theme/bin/generate-primitive-tokens/index.ts
@@ -11,9 +11,6 @@ import {
 	OKLCH,
 	sRGB,
 	type PlainColorObject,
-	// Disable reason: ESLint resolver can't handle `exports`. Import resolver
-	// checking is redundant in TypeScript files.
-	// eslint-disable-next-line import/no-unresolved
 } from 'colorjs.io/fn';
 
 /**