+++ packages/theme/src/color-ramps/test/index.test.ts
@@ -1,9 +1,6 @@
 /**
  * External dependencies
  */
-// Disable reason: ESLint resolver can't handle `exports`. Import resolver
-// checking is redundant in TypeScript files.
-// eslint-disable-next-line import/no-unresolved
 import { parse, to, serialize, sRGB } from 'colorjs.io/fn';
 
 /**