+++ packages/theme/src/use-theme-provider-styles.ts
@@ -10,9 +10,6 @@ import {
 	sRGB,
 	HSL,
 	type ColorTypes,
-	// Disable reason: ESLint resolver can't handle `exports`. Import resolver
-	// checking is redundant in TypeScript files.
-	// eslint-disable-next-line import/no-unresolved
 } from 'colorjs.io/fn';
 import memoize from 'memize';